+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- `<details>` elements in `_repr_html_` output now trigger line group resize when toggled.
+
 ## 0.5.0 - 2026-01-27
 
 ### Fixed